The main file :)
Hope you will understand the code even without reading the comments given in the code. Hope this code helps you. This will be first and the last update for this code, but I'm thinking about thinking about optimizing the code just a bit as the code stutters sometimes. Then again, I don't know if I would be able to fix those issues.